one xorcodec to rule them all
