+++ crates/client/src/workflow_handle.rs
@@ -1083,10 +1083,10 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use futures_util::{FutureExt, future::BoxFuture};
+    use crate::test_helpers::XorCodec;
     use std::collections::HashMap;
     use temporalio_common::{
-        data_converters::{DefaultFailureConverter, PayloadCodec},
+        data_converters::DefaultFailureConverter,
         protos::temporal::api::{
             common::v1::{Memo, SearchAttributes},
             enums::v1::WorkflowExecutionStatus as ProtoWorkflowExecutionStatus,
@@ -1094,35 +1094,6 @@ mod tests {
             workflow::v1::WorkflowExecutionConfig,
         },
     };
-
-    struct XorCodec;
-
-    impl PayloadCodec for XorCodec {
-        fn encode(
-            &self,
-            _context: &SerializationContextData,
-            payloads: Vec<Payload>,
-        ) -> BoxFuture<'static, Vec<Payload>> {
-            async move {
-                payloads
-                    .into_iter()
-                    .map(|mut payload| {
-                        payload.data.iter_mut().for_each(|byte| *byte ^= 0x42);
-                        payload
-                    })
-                    .collect()
-            }
-            .boxed()
-        }
-
-        fn decode(
-            &self,
-            context: &SerializationContextData,
-            payloads: Vec<Payload>,
-        ) -> BoxFuture<'static, Vec<Payload>> {
-            self.encode(context, payloads)
-        }
-    }
 
     #[tokio::test]
     async fn workflow_result_details_support_typed_decoding() {
